Create Automatic File Sorter; Python
This code is designed to sort files on a Windows computer. I utilized the shutil module to facilitate the file operations. Here's a breakdown of how it works:

Created variables to set up the necessary parameters.
Used a loop to create the required folders.
Implemented an IF, ELIF, ELSE statement to sort the files into these folders based on their tags.
Additionally, this function can be adjusted to automatically sort files at regular intervals by incorporating a timer for internal execution.