Create World Population Growth EDA in Pandas
In this project, I employed Exploratory Data Analysis (EDA) to gain insights into World Population data. I began with some basic queries to comprehend and verify the integrity of the dataset. Subsequently, I performed queries to analyze the data by continent. I then created line charts to visualize population growth over time for each continent and followed this with a box plot to identify countries that are outliers compared to the average population.